Move the debug_probe module into probe
