+++ probe-rs/src/cores/m33.rs
@@ -2,7 +2,7 @@
 //!
 
 use crate::memory::MI;
-use crate::probe::debug_probe::{DebugProbeError, MasterProbe};
+use crate::probe::{DebugProbeError, MasterProbe};
 use crate::target::{
     BasicRegisterAddresses, Core, CoreInformation, CoreRegister, CoreRegisterAddress,
 };
